+++ src/components/email-chat/email-chat.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useCurrentChatIdStore } from "../../store/current-chat-id-store";
-import { EmailChatForm } from "./email-chat-form/email-chat-form";
+import { EmailChatForm } from "./email-chat-form";
 import { GetStartedEmailChat } from "./get-started-email-chat";
 import { EmailChatMessages } from "./email-chat-messages";
 import { useChatHistoryStore } from "../../store/chat-history-store";
