+++ src/store/api.ts
@@ -4,6 +4,13 @@ import { File as ExtractedFile } from "./types";
 import { useErrorStore } from "./error-store";
 import { SYSTEM_PROMPTS } from "../system-prompts";
 
+const systemPrompts: { [key: string]: string } = {
+	"/email": SYSTEM_PROMPTS.CHAT_SYSTEM_PROMPT,
+	"/email-chat": SYSTEM_PROMPTS.EMAIL_SYSTEM_PROMPT,
+	"/note": SYSTEM_PROMPTS.NOTE_SYSTEM_PROMPT,
+	"/summary": SYSTEM_PROMPTS.SUMMARY_SYSTEM_PROMPT,
+};
+
 export async function streamChatResponse() {
 	const chatId = useCurrentChatIdStore.getState().currentChatId;
 	const { handleError } = useErrorStore.getState();
@@ -22,13 +29,6 @@ export async function streamChatResponse() {
 	}
 
 	const location = window.location.pathname;
-	const systemPrompt =
-		{
-			"/email": SYSTEM_PROMPTS.CHAT_SYSTEM_PROMPT,
-			"/email-chat": SYSTEM_PROMPTS.EMAIL_SYSTEM_PROMPT,
-			"/note": SYSTEM_PROMPTS.NOTE_SYSTEM_PROMPT,
-			"/summary": SYSTEM_PROMPTS.SUMMARY_SYSTEM_PROMPT,
-		}[location] || SYSTEM_PROMPTS.CHAT_SYSTEM_PROMPT;
 
 	const url = `${import.meta.env.VITE_API_URL}/chat`;
 
@@ -44,7 +44,8 @@ export async function streamChatResponse() {
 				messages: [
 					{
 						role: "system",
-						content: systemPrompt,
+						content:
+							systemPrompts[location] || SYSTEM_PROMPTS.CHAT_SYSTEM_PROMPT,
 					},
 					...previousMessages.map(({ role, content }) => ({
 						role,
